requirements.txt
cffi==1.15.1
install==1.3.5
Pillow==10.0.0
pycparser==2.21
pyheif==0.7.1
